requirements.txt
appdirs==1.4.3
Django==1.11
djangorestframework==3.6.2
gunicorn==19.7.1
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
requests==2.13.0
six==1.10.0
